Create 012. Integer to Roman
How to use two arrays in tandem instead of just a map. My original thought is to convert the number to a string so that we can go from the most significant to the least to get the values. But now it seems that we don't. The problem is mechanic. 

The author is quite smart to use line 11 to get the remainder of the number past the most significant number.

http://bangbingsyb.blogspot.com/2014/11/leetcode-integer-to-roman.html