Create 029. Divide Two Integers
How to take care of overflow. How to take care of sign.

The loop is basically answering this question: If the divisor is smaller or equal to the dividend, then the res must add 1 or more. The inner while loop tries to speed up things a little bit. Is it twice as large, if so, add two instead of one. 

If you just subtract the divisor from the dividend, you will have a time out error. More than two million vs 32. 

Binary search is powerful.